requirements.txt
Flask == 3.0.3
pymongo == 4.15.4
python-dotenv == 1.0.1